🐛 fix : 앱 라우터 서버 초기 데이터 캐싱 문제 해결
