+++ app/(team)/get-started/page.tsx
