--- app/_components/QureyProvider.tsx
@@ -1,18 +1,0 @@
-"use client";
-
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { useState } from "react";
-
-function QueryProvider({ children }: React.PropsWithChildren) {
-	const [queryClient] = useState(() => new QueryClient());
-
-	return (
-		<QueryClientProvider client={queryClient}>
-			{children}
-			<ReactQueryDevtools initialIsOpen={false} />
-		</QueryClientProvider>
-	);
-}
-
-export default QueryProvider;
